# Make game speed independent of fps