Before, at speeds at or above 100%, the game played at 1 (or 2 at 200%, 3 at 300%, etc...) in-game ticks per frame. This caused 100% game speed to be much faster on 144hz monitors and slower when fps was low.

This PR fixes that, and now the game always runs at the same speed regardless of fps, with a maximum limit of 20 ticks per frame.

This also effectively frameskips when fps is low, mantaining the in-game speed consistent.

There are some slight differences in speed, that aren't really noticeable on speeds below 100% but are palpable in the 200% and 300% game speeds (for the old code, the speeds are based on a 60fps framerate):

```
         |  MS per tick - avg
   Game  |--------------------
  speed  | Old code | New code
------------------------------
   10%   | 502.00   | 502.47
   20%   | 352.00   | 344.15
   30%   | 242.00   | 235.72
   40%   | 162.00   | 161.45
   50%   | 112.00   | 110.58
   60%   |  82.00   |  75.74
   70%   |  57.00   |  51.87
   80%   |  37.00   |  35.53
   90%   |  22.00   |  24.33
  100%   |  16.67   |  16.67
  200%   |   8.33   |  11.41
  300%   |   5.55   |   7.82
  400%   |   4.16   |   5.35
  500%   |   3.33   |   3.66
```