@crudelios 2020-08-04
I won't be home for the next two weeks. My monitor is also 60fps, so the way I check is to disable vsync on `platform/screen.c` and introduce a delay (using `SDL_Delay`) on every frame to simulate a certain fps.

If you want to test you can use that "technique", otherwise I'll only be able to do it in two weeks!