@attrition 2020-08-04
If you can tell me what to look for I can check what happens on high fps monitors. I have 144hz to test with.