@attrition 2020-08-04
> 
> 
> > If you can tell me what to look for I can check what happens on high fps monitors. I have 144hz to test with.
> 
> Thanks!
> Could you get [this build](https://bintray.com/bvschaik/julius-branches/windows/improve-game-speed-1.4.1.88-662bd220#files) and see how long it takes for 1 month to pass on 100% game speed? It should be around 13 seconds. On regular Julius builds, it's most likely running about twice as fast on a 144Hz monitor.
> Also, going from 90% to 100% should be a slight increase in speed, just like going from 80% to 90%, not a big jump, like it probably is now.

13 seconds @ 144hz @ 100%. 90 to 100% was a small uptick in speed as expected.