@attrition 2020-08-04
Also as a sanity check, the latest stable build of Julius passed a month in about 5-6 seconds at 100% at 144hz.