@crudelios 2020-08-05
Can you also test/compare at 200% please? @bvschaik what time do you get on your 60fps monitor to compare? If it checks out I'll close tbis PR so you can add your changes.