@bvschaik 2020-08-05
I actually tweaked the speeds for > 100% slightly in my branch since they will probably only be used to speed up the game significantly (for testing those keep-em-stable-for-a-century cities).

The following table lists the current code's calculated speeds at 60 and 144Hz with vsync enabled. This is assuming the computer is powerful enough to run 1 game tick within one screen refresh. I measured the ones for 100 and 200% on my 60Hz screen and they match up. The last column is what my new code does, regardless of screen refresh rate.

    Game  |     Month (seconds)
    Speed | 60Hz  | 144Hz | New code
    -------------------------------
     100% | 13.33 |  5.56 | 12.80
     200% |  6.67 |  2.78 |  6.40
     300% |  4.44 |  1.85 |  4.00
     400% |  3.33 |  1.39 |  2.40
     500% |  2.67 |  1.11 |  1.60