@attrition 2020-08-05
@ 144hz:
400% was just shy of 3 seconds and 500% was just about 2 seconds. It looks to be working correctly from your given chart.